feat: adding two new commands to help with number removal from slugs
Embarcadero asked us to ensure that slugs do not have a number appended to the end (e.g. `<site>/news/2025/03/10/something-important-happened-2`, they want to make sure it's `<site>/news/2025/03/10/something-important-happened`. The first command will process all posts and remove the appended number where it can. It will also produce a QA file for a review of all the posts that were and weren't updated.