add: New Windy City Migrator to fix MP3 URLs in Queercast posts.
The folder structure for the site Windy City previously came from was a little different from ours. They provided us all their MP3s, so this was just a matter of finding the MP3 (using the indicated URL in the post in question), then copying from that tranche over to the new folder within our folder structure (`wp-content/uploads/YYYY/MM/*`). Finally, the URL needed to be updated so that the new file location was reflected.