fix: issue where EOF is reached prematurely, and `get_row` returns bool
There are some CSV files that have an extra line item at the end of it. It's empty though. But this means that `feof` returns `false` though. This then then leads to trying to get this row with `get_row` and the `fgetcsv` call returning `false` instead of an expected array.

These changes will now let `get_row` to return `bool` (which can happen when calling `fgetcsv`), and when a `bool` is received, it means we have reached the end of file, there are no more rows to process, so we should either return/exit out of the current function or throw an Exception.