fix: ensure that yoast primary category can be nullified
We don't remove any pre-existing postmeta. So previously, if `_yoast_wpseo_primary_category` already existed, it would just remain as is, because there was no new value to update it with. This line ensures that the value is reset. Further down in the migrator, if there is a need for a primary category, that meta is then updated to match the term_id of the intended category.